backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.2
psycopg2-binary==2.9.9
Pillow==10.4.0
gunicorn==22.0.0
PyJWT==2.9.0
